127 ci caching (#195)
* wip- first try at caching pip and docker images

* added pip caching; added docker caching; modified install/setup scripts to optionall skip image builds

* appeased the pylint gods

* fixed import order

* pass in github token

+++ install_linux.sh
@@ -14,7 +14,7 @@ db_pw=''
 api_url=''
 metrics_url=''
 
-while getopts "p:a:m:" o; do
+while getopts "p:a:m:n" o; do
     case "$o" in
         p)
             db_pw=${OPTARG}
@@ -24,6 +24,9 @@ while getopts "p:a:m:" o; do
             ;;
         m)
             metrics_url=${OPTARG}
+            ;;
+        n)
+            no_build=true
             ;;
     esac
 done
@@ -119,9 +122,11 @@ if [[ ${host_metrics_url} == *".green-coding.example"* ]];then
     fi
 fi
 
-print_message "Building / Updating docker containers"
-docker compose -f docker/compose.yml down
-docker compose -f docker/compose.yml build
+if [[ $no_build != true ]] ; then
+    echo "Building / Updating docker containers"
+    docker compose -f docker/compose.yml down
+    docker compose -f docker/compose.yml build
+fi
 
 echo ""
 echo -e "${GREEN}Successfully installed Green Metrics Tool!${NC}"
